requirements.txt
Flask==2.3.3
openai==1.13.3
python-dotenv==1.0.1
